factor out push sender/receive and improve push parsing
Sender and receiver information are factored into their own datatypes,
which are included in `Push 'Existing` values. These are parsed in such
a way that if *all* their fields are present, we get one such sender or
receiver value. The sender value is mandatory, and is either a client
(user) or a channel. User senders include additional fields such as
email address. Channels include only the channel ID and their name.

This fixes parsing of channel pushes.